requirements.local.txt
factory_boy
faker
flake8
flask-debugtoolbar
flask-testing
ptpython
